error

+++ code/python/acquire_pubs.py
@@ -138,6 +138,10 @@ def make_json_folder():
     """
 
     """
+    task_name = 'make_json_folder'
+    if work_to_do(task_name): continue
+    work_completed(task_name, 0)
+
     shutil.rmtree(os.path.join(retrieve_path('pub_json')))
 
     # list directories with json
@@ -197,6 +201,7 @@ def make_json_folder():
             df.to_csv(os.path.join(retrieve_path('pub_links')))
             print('pubs founds: ' + str(len(links)))
 
+    work_completed(task_name, 1)
 
 def search_meta():
     """
@@ -217,10 +222,10 @@ def search_term():
 
     task_name = 'search_pubs'
     if work_to_do(task_name):
-        work_completed('search_pubs', 0)
+        work_completed(task_name, 0)
         search_crossref()
         search_gscholar()
-        work_completed('search_pubs', 1)
+        work_completed(task_name, 1)
 
 
 def search_crossref():
